Provisioning: Fix folder path for dashboards
